refactor NodeProvider
