+++ src/containers/Elements/index.js
@@ -1,3 +1,4 @@
+export { default as NodePanel } from './NodePanel';
 export { default as NodeProvider } from './NodeProvider';
 export { default as NodeProviderPanels } from './NodeProviderPanels';
 export { default as PromptSwiper } from './PromptSwiper';
